backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
passlib[bcrypt]
PyJWT
python-jose
email-validator
pymongo
motor
pinecone-client
transformers
sentence-transformers
torch
torchvision
torchaudio
scikit-learn
pandas
numpy
scipy
tqdm
requests
loguru